+++ alarmer/provider/apprise.py
@@ -18,5 +18,7 @@ class AppriseProvider(Provider):
         self.app = apprise.Apprise(services, **(apprise_options or {}))
         self.notify_options = notify_options or {}
 
-    def send(self, message: str, exc: Optional[BaseException] = None):
+    def send(
+        self, message: str, exc: Optional[BaseException] = None, context: Optional[dict] = None
+    ):
         return self.app.notify(body=message, **self.notify_options)