+++ alarmer/provider/feishu.py
@@ -14,7 +14,9 @@ class FeiShuProvider(Provider):
         super().__init__()
         self.webhook_url = webhook_url
 
-    def send(self, message: str, exc: Optional[BaseException] = None):
+    def send(
+        self, message: str, exc: Optional[BaseException] = None, context: Optional[dict] = None
+    ):
         data = {"msg_type": "text", "content": {"text": message}}
         return requests.post(
             self.webhook_url,